fix: 修复PSStreamDemuxer Pack Header识别失败的类型比较bug
PACK_START_CODE_ID和SYSTEM_HEADER_ID从byte改为int类型。
原byte类型(0xBA=-70)与int类型(streamId&0xFF=186)比较时因符号扩展导致186==-70为false，
Pack Header被错误当作非视频PES跳过，后续所有视频PES无法解析。

Coze-Commit-Type: user
Coze-User-ID: 1751938984068412
Coze-Conversation-ID: 8983462

+++ iot-platform/iot-protocol/iot-protocol-haikang-isup/src/main/java/org/springblade/modules/iot/haikangisup/handler/PSStreamDemuxer.java
@@ -15,9 +15,9 @@ import java.util.List;
 public class PSStreamDemuxer {
     // ===================== PS 标准起始码常量 =====================
     /** Pack起始码 0x000001BA */
-    private static final byte PACK_START_CODE_ID = (byte) 0xBA;
+    private static final int PACK_START_CODE_ID = 0xBA;
     /** System Header起始码 0x000001BB */
-    private static final byte SYSTEM_HEADER_ID = (byte) 0xBB;
+    private static final int SYSTEM_HEADER_ID = 0xBB;
     /** PES公共前缀 00 00 01 */
     private static final int PES_PREFIX = 0x000001;
     /** PES流ID掩码 */
